refactor: extract config-resolver and helpers from agent/executor.ts
Decompose executor.ts (898 → 691 lines) into focused modules:
- config-resolver.ts: multi-source config merging (definition + DB + YAML + myco.yaml)
- ollama-context.ts: Ollama model variant creation for context window sizing
- wave-computation.ts: Kahn's algorithm topological sort + phase session IDs
