+++ src/agent/ollama-context.ts
@@ -1,0 +1,55 @@
+/**
+ * Ollama model context window management.
+ *
+ * Creates Ollama model variants with a specific context length by generating
+ * a Modelfile and running `ollama create`. The Anthropic-compatible endpoint
+ * (/v1/messages) always loads models at default context — this is the only
+ * reliable way to override it.
+ */
+
+/** Timeout for Ollama model pre-load request (ms). */
+const OLLAMA_PRELOAD_TIMEOUT_MS = 30_000;
+
+/**
+ * Ensure an Ollama model variant exists with the desired context length.
+ *
+ * The Anthropic-compatible endpoint (/v1/messages) always loads models at
+ * default context — it ignores /api/chat preloads and API-created params.
+ * The only reliable way is `ollama create` with a Modelfile containing
+ * `PARAMETER num_ctx`. Creates a variant named `{model}-ctx{contextLength}`.
+ *
+ * Returns the variant model name to use.
+ */
+export async function ensureOllamaContextVariant(
+  model: string,
+  contextLength: number,
+): Promise<string> {
+  const { execFileSync } = await import('node:child_process');
+  const { writeFileSync, unlinkSync } = await import('node:fs');
+  const { tmpdir } = await import('node:os');
+  const { join } = await import('node:path');
+
+  const baseName = model.replace(/:latest$/, '');
+  const variantName = `${baseName}-ctx${contextLength}`;
+
+  try {
+    // Check if variant already exists
+    execFileSync('ollama', ['show', variantName], { stdio: 'ignore' });
+    return variantName;
+  } catch {
+    // Doesn't exist — create it
+  }
+
+  try {
+    const modelfilePath = join(tmpdir(), `myco-modelfile-${Date.now()}`);
+    writeFileSync(modelfilePath, `FROM ${model}\nPARAMETER num_ctx ${contextLength}\n`);
+    execFileSync('ollama', ['create', variantName, '-f', modelfilePath], {
+      stdio: 'ignore',
+      timeout: OLLAMA_PRELOAD_TIMEOUT_MS,
+    });
+    try { unlinkSync(modelfilePath); } catch { /* cleanup best-effort */ }
+    return variantName;
+  } catch {
+    return model; // Fall back to original
+  }
+}
